handling exception in deleteCard

+++ src/flashcards/Records.java
@@ -3,7 +3,7 @@ package flashcards;
 import java.util.ArrayList;
 import flashcards.cards.Card;
 
-public class Records implements java.io.Serializable{
+public class Records implements java.io.Serializable {
     private ArrayList<Category> categories;
     private ArrayList<Card> cards;
     private ArrayList<Deck> decks;
@@ -23,18 +23,21 @@ public class Records implements java.io.Serializable{
     }
 
     public void deleteCard(Card card, User user) throws InvalidUserException {
-        // Also check if user is the owner
-
-        int index = this.cards.indexOf(card);
-        if(index != -1) {
+        if(card.getCreator().equals(user)){
+            int index = this.cards.indexOf(card);
+            if(index != -1) {
             cards.remove(index);
-        } // Else throw exception
+            } // Else throw exception
 
-        // Deleting the card from all the decks as well!
+            // Deleting the card from all the decks as well!
 
-        for(Deck deck: decks) {
-            deck.deleteCard(card, user);
+            for(Deck deck: decks) {
+                deck.deleteCard(card, user);
+            }
+        } else {
+            throw new InvalidUserException("User is not the owner of the card");
         }
+
     }
     public void addDeck(Deck deck) {
         this.decks.add(deck);
